cv-dcm-ingestor/requirement.txt
requests
mysql-connector-python
pydicom
Pillow
numpy
pandas